feat: NftCardWithEffects component created
