+++ components/base/NftCard/index.tsx
@@ -1,1 +1,2 @@
-export { default } from "./NftCard";
+export { default } from './NftCard';
+export { default as NftCardWithEffects } from './NftCardWithEffects';
